server/requirements.txt
flask
transformers
opencv-python
pillow
numpy
torch